Move helper functions out of app into utils and pdf_utils
